moved ratings, bookmark functions
